Added better command parsing to handle noise. If you add a < to the beginning of each command (added to android code) and a > to the end, it will disregard every byte outside of it which stops most of the unrecognised problems.
Added a < and > to the serialReply so for better handling within the python script.

Added two extra commands. l[number] and r[number] which writes [number] to the servo. It also sets servosForcedActive to true which tells checkIfStopBot() to keep going unless its not safeToProceed(). Dont include the [] in the command.